Image offline augmentation
Image augmentation is a way of creating synthetic data. Offline augmentation means that the synthetic data is done outside the algorithm, and the data is then saved as a physical copy. I have managed to multiply my datasets of L,M, and N, by 10. I now have around ~1150 of each images.